Fix issue with NAG (#1939)

+++ src/io/buffer/buffer_4d_npar.F90
@@ -44,7 +44,7 @@ module buffer_4d_npar
   integer, private :: npar
 
   type, extends(buffer_t) :: buffer_4d_npar_t
-     integer(kind=8), dimension(5) :: shape_dims, start_dims, count_dims
+     integer(kind=i8), dimension(5) :: shape_dims, start_dims, count_dims
      real(kind=dp), private, allocatable :: data_dp(:,:,:,:,:)
      real(kind=sp), private, allocatable :: data_sp(:,:,:,:,:)
    contains
